Basic Movement
The asset pack has been set to use. A simple map's been created. Inside is a space fighter Player character. Currently it uses basic Rigidbody controls for movement; but the problems are already showing up. The rotation has to be fixed so the character doesn't fall down, but this is sure to invoke future problems. Better movement options should be considered in all games.